refactor: cập nhật chỉ số lên int

+++ src/main/java/com/teamobi/mobiarmy2/fight/Boss.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public abstract class Boss extends Player {
     private String name;
 
-    public Boss(FightManager fightManager, byte characterId, String name, short x, short y, short width, short height, short maxHp, int xpExist) {
+    public Boss(FightManager fightManager, byte characterId, String name, short x, short y, short width, short height, int maxHp, int xpExist) {
         super(fightManager, characterId, x, y, width, height, maxHp, xpExist);
         this.name = name;
         this.stamina = 100;
