+++ src/main/java/com/teamobi/mobiarmy2/fight/boss/Balloon.java
@@ -21,7 +21,7 @@ public class Balloon extends Boss {
     public void turnAction() {
         // Thêm Balloon Eye nếu bodyParts[4] là null, Gun và Gun Big bị tiêu diệt
         if (bodyParts[4] == null && bodyParts[1].isDead() && bodyParts[2].isDead()) {
-            BalloonEye balloonEye = new BalloonEye(fightManager, (short) (x + 55), (short) (y - 27), (short) 1000);
+            BalloonEye balloonEye = new BalloonEye(fightManager, (short) (x + 55), (short) (y - 27), 1000);
             bodyParts[4] = balloonEye;
             fightManager.addPendingBoss(balloonEye);
             fightManager.doNextTurn();
