+++ src/main/java/com/teamobi/mobiarmy2/fight/boss/BalloonFanBack.java
@@ -4,7 +4,7 @@ import com.teamobi.mobiarmy2.fight.Boss;
 import com.teamobi.mobiarmy2.fight.FightManager;
 
 public class BalloonFanBack extends Boss {
-    public BalloonFanBack(FightManager fightManager, short x, short y, short maxHp) {
+    public BalloonFanBack(FightManager fightManager, short x, short y, int maxHp) {
         super(fightManager, (byte) 20, "Balloon Fan Back", x, y, (short) 10, (short) 19, maxHp, 4);
         super.isFlying = true;
     }
