+++ src/main/java/com/teamobi/mobiarmy2/fight/boss/BigBoom.java
@@ -6,7 +6,7 @@ import com.teamobi.mobiarmy2.fight.Player;
 import com.teamobi.mobiarmy2.util.RandomUtil;
 
 public class BigBoom extends Boss {
-    public BigBoom(FightManager fightManager, short x, short y, short maxHp) {
+    public BigBoom(FightManager fightManager, short x, short y, int maxHp) {
         super(fightManager, (byte) 12, "Big boom", x, y, (short) 28, (short) 28, maxHp, 4);
     }
