+++ src/main/java/com/teamobi/mobiarmy2/fight/boss/SmallBoom.java
@@ -5,7 +5,7 @@ import com.teamobi.mobiarmy2.fight.FightManager;
 import com.teamobi.mobiarmy2.fight.Player;
 
 public class SmallBoom extends Boss {
-    public SmallBoom(FightManager fightManager, short x, short y, short maxHp) {
+    public SmallBoom(FightManager fightManager, short x, short y, int maxHp) {
         super(fightManager, (byte) 11, "Small Boom", x, y, (short) 18, (short) 18, maxHp, 2);
     }
