+++ src/main/java/com/teamobi/mobiarmy2/fight/boss/UFOPet.java
@@ -6,7 +6,7 @@ import com.teamobi.mobiarmy2.fight.FightManager;
 
 public class UFOPet extends Boss {
 
-    public UFOPet(FightManager fightManager, short x, short y, short maxHp) {
+    public UFOPet(FightManager fightManager, short x, short y, int maxHp) {
         super(fightManager, (byte) 16, "UFO", x, y, (short) 51, (short) 46, maxHp, 4);
         super.isFlying = true;
     }
